Move QR decoder plugin to shared package
