--- mobile/apps/auth/plugins/qr/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'ente_auth_qr'
