--- mobile/apps/auth/plugins/qr/pubspec.yaml
@@ -1,29 +1,0 @@
-name: ente_auth_qr
-description: "A QR code reader plugin for Ente."
-version: 0.0.1
-homepage: https://ente.com
-
-environment:
-  sdk: ">=3.10.0 <4.0.0"
-  flutter: '>=3.3.0'
-
-dependencies:
-  flutter:
-    sdk: flutter
-  plugin_platform_interface: 2.1.8
-
-dev_dependencies:
-  flutter_lints: 6.0.0
-  flutter_test:
-    sdk: flutter
-
-# The following section is specific to Flutter packages.
-flutter:
-  # This section identifies this Flutter project as a plugin project.
-  plugin:
-    platforms:
-      android:
-        package: io.ente.auth.ente_auth_qr
-        pluginClass: EnteQrPlugin
-      ios:
-        pluginClass: EnteQrPlugin
