+++ mobile/apps/auth/pubspec.yaml
@@ -27,8 +27,6 @@ dependencies:
   email_validator: 3.0.0
   ente_accounts:
     path: ../../packages/accounts
-  ente_auth_qr:
-    path: plugins/qr
   ente_base:
     path: ../../packages/base
   ente_configuration:
@@ -47,6 +45,8 @@ dependencies:
     path: ../../packages/network
   ente_pure_utils:
     path: ../../packages/ente_pure_utils
+  ente_qr:
+    path: ../../packages/ente_qr
   ente_strings:
     path: ../../packages/strings
   ente_ui:
