+++ mobile/apps/photos/android/build.gradle
@@ -8,6 +8,7 @@ allprojects {
         google()
         jcenter()
         mavenCentral()
+        maven { url 'https://jitpack.io' }
         // mavenLocal() // for FDroid
         maven {
             url "${project(':ffmpeg_kit_flutter').projectDir}/libs"
