--- mobile/apps/photos/plugins/ente_qr/CHANGELOG.md
@@ -1,8 +1,0 @@
-## 0.0.1
-
-- Initial release of ente_qr plugin
-- Support for scanning QR codes from image files
-- Android implementation using ZXing library
-- iOS implementation using Core Image framework
-- Comprehensive error handling
-- Example app demonstrating usage with file picker
