+++ mobile/apps/photos/pubspec.yaml
@@ -71,7 +71,7 @@ dependencies:
   ente_pure_utils:
     path: ../../packages/ente_pure_utils
   ente_qr:
-    path: plugins/ente_qr
+    path: ../../packages/ente_qr
   ente_qr_ui:
     path: ../../packages/qr
   ente_rust:
