+++ mobile/packages/ente_qr/.metadata
