+++ mobile/packages/ente_qr/CHANGELOG.md
@@ -1,0 +1,8 @@
+## 0.0.1
+
+- Initial release of ente_qr plugin
+- Support for scanning QR codes from image files
+- Android implementation using zxing-cpp
+- iOS implementation using Vision
+- Comprehensive error handling
+- Example app demonstrating usage with file picker
