+++ mobile/packages/ente_qr/analysis_options.yaml
