+++ mobile/packages/ente_qr/android/settings.gradle
