+++ mobile/packages/ente_qr/android/src/main/AndroidManifest.xml
