+++ mobile/packages/ente_qr/ios/Classes/EnteQrPlugin.swift
