+++ mobile/packages/ente_qr/lib/ente_qr_method_channel.dart
