+++ mobile/packages/ente_qr/lib/ente_qr_platform_interface.dart
