+++ mobile/packages/ente_qr/lib/src/desktop_qr_platform.dart
@@ -1,0 +1,2 @@
+export 'desktop_qr_platform_stub.dart'
+    if (dart.library.io) 'desktop_qr_platform_io.dart';
