+++ mobile/packages/ente_qr/lib/src/desktop_qr_platform_stub.dart
@@ -1,0 +1,3 @@
+import 'package:ente_qr/ente_qr_platform_interface.dart';
+
+EnteQrPlatform? desktopQrPlatform() => null;
